@@ -14,26 +14,26 @@
     }
   ],
   "require": {
-    "php": "^8.1",
-    "typo3/cms-core": "^12.4",
-    "typo3/cms-extbase": "^12.4",
-    "typo3/cms-dashboard": "^12.4",
-    "ssch/t3-messenger": "^2.0",
+    "php": "^8.2",
+    "typo3/cms-core": "^13.4",
+    "typo3/cms-extbase": "^13.4",
+    "typo3/cms-dashboard": "^13.4",
+    "ssch/t3-messenger": "^3.0",
     "ext-json": "*"
   },
   "require-dev": {
     "symplify/easy-coding-standard": "^12.0",
     "php-parallel-lint/php-parallel-lint": "^1.3",
-    "phpstan/phpstan": "^1.8",
-    "phpstan/phpstan-phpunit": "^1.2",
-    "phpstan/phpstan-strict-rules": "^1.4",
-    "saschaegerer/phpstan-typo3": "^1.8",
-    "phpstan/extension-installer": "^1.2",
-    "typo3/testing-framework": "^8.0.9",
+    "phpstan/phpstan": "^2.1.33",
+    "phpstan/phpstan-phpunit": "^2.0.11",
+    "phpstan/phpstan-strict-rules": "^2.0.7",
+    "saschaegerer/phpstan-typo3": "^2.1.1",
+    "phpstan/extension-installer": "^1.4",
+    "typo3/testing-framework": "^9.3.0",
     "helhum/typo3-console": "^8.0",
-    "typo3/minimal": "^12.4",
-    "typo3/cms-lowlevel": "^12.4",
-    "rector/rector": "^1.0.3"
+    "typo3/minimal": "^13.4",
+    "typo3/cms-lowlevel": "^13.4",
+    "rector/rector": "^2.3.1"
   },
   "replace": {
     "typo3-ter/t3-messenger-dashboard": "self.version"
@@ -84,7 +84,7 @@
   },
   "extra": {
     "branch-alias": {
-      "dev-main": "2.0.x-dev"
+      "dev-main": "3.0.x-dev"
     },
     "typo3/cms": {
       "extension-key": "t3_messenger_dashboard",
