@@ -18,11 +18,11 @@ use Ssch\T3MessengerDashboard\Repository\FailedMessageRepository;
 use TYPO3\CMS\Core\Http\JsonResponse;
 use TYPO3\CMS\Core\SingletonInterface;
 
-final class FailedMessageController implements SingletonInterface
+final readonly class FailedMessageController implements SingletonInterface
 {
     public function __construct(
-        private readonly FailedMessageRepository $failedMessageRepository,
-        private readonly JsonSerializer $jsonSerializer
+        private FailedMessageRepository $failedMessageRepository,
+        private JsonSerializer $jsonSerializer
     ) {
     }
 
