@@ -1,16 +1,5 @@
 <?php
 
-call_user_func(static function () {
-    \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addTypoScriptSetup(
-        "@import 'EXT:t3_messenger_dashboard/Configuration/TypoScript/setup.typoscript'"
-    );
-
-    $iconRegistry = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(\TYPO3\CMS\Core\Imaging\IconRegistry::class);
-    $iconRegistry->registerIcon(
-        'tx-messenger-failed-messages-icon',
-        \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
-        [
-            'source' => 'EXT:t3_messenger_dashboard/Resources/Public/Icons/failed-messages.svg',
-        ]
-    );
-});
+\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addTypoScriptSetup(
+    "@import 'EXT:t3_messenger_dashboard/Configuration/TypoScript/setup.typoscript'"
+);
